+++ project/urls.py
@@ -1,18 +1,3 @@
-"""cradle_of_mankind URL Configuration
-
-The `urlpatterns` list routes URLs to views. For more information please see:
-    https://docs.djangoproject.com/en/3.1/topics/http/urls/
-Examples:
-Function views
-    1. Add an import:  from my_app import views
-    2. Add a URL to urlpatterns:  path('', views.home, name='home')
-Class-based views
-    1. Add an import:  from other_app.views import Home
-    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
-Including another URLconf
-    1. Import the include() function: from django.urls import include, path
-    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
-"""
 from django.contrib import admin
 from django.conf import settings
 from django.conf.urls import url
@@ -20,7 +5,7 @@ from django.conf.urls.static import static
 from django.urls import path
 from django.urls.conf import include
 from users.views import index
-from .views import health, upload, parse, table_to_dataframe
+from .views import health, upload, table_to_dataframe
 # from .extract_table import table_to_dataframe
 
 urlpatterns = [
@@ -30,10 +15,12 @@ urlpatterns = [
     path('quality_control/', include('quality_control.urls')),
     path('masterdata/', include('masterdata.urls')),
     path('', index, name='index'),
+    path ('form/', table_to_dataframe, name="form"),
     path('table/', table_to_dataframe, name="table"),
     url(r'^health$', health),
     path('upload/', upload, name='upload'),
-    path('parse/', parse, name='parse'),
+    path('parse/', include('spacy_parse.urls')),
+    path('', index, name='index'),
 ] + static(settings.STATIC_URL, document_root=settings.STATIC_ROOT)
 
 if settings.DEBUG:
