+++ requirements.txt
@@ -11,7 +11,6 @@ django-admin==2.0.1
 django-allauth==0.44.0
 django-debug-toolbar==3.2
 django-crispy-forms==1.9.2
-django-pandas==0.14.1
 django-simple-history==2.12.0
 django-sslserver==0.22
 django-userforeignkey==0.3.0
@@ -38,5 +37,6 @@ toml==0.10.2
 urllib3==1.26.2
 whitenoise==5.1.0
 spacy==3.0.0
+https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0.tar.gz#egg=en_core_web_sm
