--- table_to_csv.csv
@@ -1,29 +1,0 @@
-"                                       Dataset BEAUTY              GAMES              STEAM              ML-1M
-0                            Metric Recall@10 NDCG@10  Recall@10 NDCG@10  Recall@10 NDCG@10  Recall@10 NDCG@10
-1                            (A) POPRec 0.4003 0.2277      0.4724 0.2779      0.7172 0.4535      0.4329 0.2377
-2                               (B) BPR 0.3775 0.2183      0.4853 0.2875      0.7061 0.4436      0.5781 0.3287
-3                               (C) FMC 0.3771 0.2477      0.6358 0.4456      0.7731 0.5193      0.6983 0.4676
-4                              (D) FPMC 0.4310 0.2891      0.6802 0.4680      0.7710 0.5011      0.7599 0.5176
-5                          (E) TRANSREC 0.4607 0.3020      0.6838 0.4557      0.7624 0.4852      0.6413 0.3969
-6                           (F) GRU4REC 0.2125 0.1203      0.2938 0.1837      0.4190 0.2691      0.5581 0.3381
-7                             (G) STAMP 0.4607 0.3020      0.6838 0.4557      0.7624 0.4852      0.6413 0.3969
-8                          (H) GRU4REC+ 0.3949 0.2556      0.6599 0.4759      0.8018 0.5595      0.7501 0.5513
-9                             (I) CASER 0.4264 0.2547      0.5282 0.3214      0.7874 0.5381      0.7886 0.5538
-10                           (J) SASREC 0.4837 0.3220      0.7434 0.5401      0.8732 0.6293      0.8233 0.5936
-11                              (K) HGN 0.4469 0.2994      0.7164 0.5209      0.7426 0.4871      0.7584 0.5241
-12                       (L) SSE-SASREC 0.4878 0.3342      0.7517 0.5535      0.8697 0.6333      0.8230 0.5995
-13                               (M) PT 0.3954 0.2449      0.6427 0.4434      0.7535 0.4853      0.7658 0.5241
-14                           (N) SSE-PT 0.5028 0.3370      0.7757 0.5660      0.8772 0.6378      0.8341 0.6281
-15                                   (O) SSE-PT++ – –                – –                – –      0.8389 0.6292
-16  2: Comparing SASRec, SSE-PT and SSE-PT++ on Mo...                NaN                NaN                NaN
-17                                sion of embeddings.                NaN                NaN                NaN
-18                                    Methods NDCG@10  Recall@10 Max Len  user dim item dim                NaN
-19                                      SASREC 0.5769         0.8045 100            N/A 100                NaN
-20                                      SASREC 0.5936         0.8233 200             N/A 50                NaN
-21                                      SASREC 0.5919         0.8202 200            N/A 100                NaN
-22                                      SSE-PT 0.6142         0.8212 100             50 100                NaN
-23                                      SSE-PT 0.6191         0.8358 200              50 50                NaN
-24                                      SSE-PT 0.6281         0.8341 200             50 100                NaN
-25                                    SSE-PT++ 0.6186         0.8318 100             50 100                NaN
-26                                    SSE-PT++ 0.6208         0.8358 200              50 50                NaN
-27                                    SSE-PT++ 0.6292         0.8389 200             50 100                NaN"
